create handleCocktailSearch
merge handleInputSearch and handleLetterClick functions into handleCocktailsSearch function
